# Avoid py25519 dependency with cryptography.