--- a/requirements.txt
+++ b/requirements.txt
@@ -1,5 +1,5 @@
 zeroconf
 gmpy2
-py25519
 hkdf
 ed25519
+cryptography