--- a/setup.py
+++ b/setup.py
@@ -41,9 +41,9 @@ setuptools.setup(
     install_requires=[
         'zeroconf',
         'gmpy2',
-        'py25519',
         'hkdf',
         'ed25519',
+        'cryptography',
     ],
     license='Apache License 2.0',
     long_description=long_description,