This is an alternative implementation of the curve 25519 key exchanged based on the cryptography library. The motivation is that cryptography is a more actively maintained dependency, works on macOS and is easier to install (ships as a wheel on most platforms).

See https://cryptography.io/en/stable/.

The cryptography library has a large community with active team-based maintenance doing regular releases. Heavy lifting is done by openssl, and a lot of work has been put into packaging. On most python installs (e.g. ones with whl support) a compiler isn't needed. It has broad platform support.

In comparison, py25519 has not had a release since 2016, does not work on macOS, and runs its tests every time you import its module.

This will fix #52 and will likely be required for #63.

It will also help make getting my HA tests running on their travis a bit easier too =)