@Jc2k 2019-01-19
# Performance testing round 1

The accessory under test is a demoserver.py lightbulb running on my Macbook (it has both of my crypto PR's applied + the fix to make demoserver.py work with ios devices).

The client is a 2011 Raspberry Pi running at 700MhZ. It's running the latest raspbian at the time of writing. I will be running the test script with this PR and without it.

The test script is:

```
import time

from homekit.controller import Controller
from homekit.controller import Session


class Timer:
    def __enter__(self):
        self.start = time.clock()
        return self

    def __exit__(self, *args):
        self.end = time.clock()
        self.interval = self.end - self.start


def test_cycle(c):
    print('Starting pair')

    with Timer() as t:
        c.perform_pairing('test', '14:00:00:00:00:02', '031-45-154')
    print('PAIR', t.interval)

    pairing = c.get_pairings()['test']

    for i in range(10):
        with Timer() as t:
            pairing.session = Session(pairing.pairing_data)
        print("VERIFY", t.interval)
        for j in range(2):
            with Timer() as t:
                pairing.list_accessories_and_characteristics()
            print("READ", t.interval)
        pairing.session.close()
        pairing.session = None

    with Timer() as t:
        c.remove_pairing('test')
    print("UNPAIR", t.interval)



if __name__ == '__main__':
    c = Controller()
    for i in range(10):
        test_cycle(c)
```

It will pair and unpair 10 times, each time it pairs it will start 10 sessions, and for each session it will call  list_accessories_and_characteristics() twice.

On my PR there is surprisingly little variance and this is a fairly representative sample:

```
PAIR 2.2934879999999964
VERIFY 0.27053300000000036
READ 0.7561139999999966
READ 0.7530199999999923
VERIFY 0.4756770000000188
READ 0.7364970000000142
READ 0.739137999999997
VERIFY 0.2681530000000123
READ 0.7481520000000046
READ 0.7537090000000148
VERIFY 0.2733340000000055
READ 0.7730669999999975
READ 0.7581849999999974
VERIFY 0.2787099999999896
READ 0.8053769999999929
READ 0.802552999999989
VERIFY 0.2788539999999955
READ 0.80855600000001
READ 0.7939010000000053
VERIFY 0.27552599999998506
READ 0.7967960000000005
READ 0.7952250000000163
VERIFY 0.2776700000000005
READ 0.8022119999999973
READ 0.8005469999999946
VERIFY 0.2758620000000178
READ 0.7998350000000016
READ 0.8032780000000059
VERIFY 0.2774650000000065
READ 0.8013090000000034
READ 0.7896429999999839
UNPAIR 0.4867059999999981
```

Pairing takes 2.27s to 2.37s, establishing a secure session takes 0.2s to 0.3s, unpair takes 0.4-0.5s (this includes establishing a final secure session). Generating the list of services and characteristics and (de)serializing it and (de)encrypting it is noticeably slower than just unpairing (both happen over secure sessions). 

The Pi is so slow that just starting python takes longer than pairing, verifying and reading. I didn't use a stopwatch, but it felt like about 15s before it started to run any code. So the CLI tools have an overhead that eclipses the impact of the crypto changes.

Comparing this against the same code without this PR  + same setup on macOS, we expect to see no change to READ times. A 25519 key exchange only happens in pair-verify - so thats where we are most expecting to see changs. A few sign/verify steps were swapped from py25519 to ed25519 (an existing dep) so we need to watch out for those too.

Again - variance is quite low. Picking a sample at random, the values on `master` are

```
Starting pair
PAIR 2.482374
VERIFY 0.29711900000000036
READ 0.7762670000000007
READ 0.7846499999999992
VERIFY 0.2713839999999994
READ 0.7827940000000009
READ 0.771605000000001
VERIFY 0.2717119999999973
READ 0.7741760000000006
READ 0.7737160000000003
VERIFY 0.27158399999999716
READ 0.7795139999999989
READ 0.7742289999999983
VERIFY 0.490351000000004
READ 0.7952689999999905
READ 0.7880839999999978
VERIFY 0.2775040000000004
READ 0.7971320000000013
READ 0.7843899999999984
VERIFY 0.2738940000000021
READ 0.7902869999999993
READ 0.7797940000000025
VERIFY 0.27266600000000096
READ 0.7683019999999985
READ 0.7644260000000003
VERIFY 0.2695469999999993
READ 0.761610000000001
READ 0.7629310000000018
VERIFY 0.27122200000000163
READ 0.7620429999999985
READ 0.7647859999999973
UNPAIR 0.4760530000000003
```

Pairing is take 2.28s - 2.48s. The PR and master are close enough for it to be a measurement error.

On master a verify takes 0.2s to .49s, but mostly it was wabout 0.27s. It's broadly un changed between my PR and master, with my PR having a tiny edge if anything.

Both branches are firmly in the 0.7s-0.8s range for reading the services and chars.

Unpairing takes 0.47s-0.50s. Again, master and this PR are comparable.