@Jc2k 2019-01-19
# Performance testing round 2 (pi as accessory)

Same tooling as before but test.py is now executed on my mac, and we run demoserver.py on master and my PR. It's here that homekit is more likely to be useful on an old pi or underpowered pi - we can stomach the 15s start up times more easily.

## master

The timings are massively different this time. With a modern Mac as a client and a 2011 700MhZ Pi as server we get:

```
PAIR 0.20615000000000006
VERIFY 0.006642000000000259
READ 0.014349999999999863
READ 0.014185999999999588
VERIFY 0.0066070000000002516
READ 0.014327999999999896
READ 0.014135000000000009
VERIFY 0.006568999999999825
READ 0.014251000000000236
READ 0.01430700000000007
VERIFY 0.006663999999999781
READ 0.01422699999999999
READ 0.014114999999999878
VERIFY 0.006775999999999893
READ 0.014129999999999754
READ 0.014171000000000156
VERIFY 0.0066390000000002836
READ 0.014264999999999972
READ 0.014202999999999744
VERIFY 0.006650000000000045
READ 0.014277000000000317
READ 0.014480999999999966
VERIFY 0.006556999999999924
READ 0.014438000000000173
READ 0.014142999999999795
VERIFY 0.006588999999999956
READ 0.014287999999999634
READ 0.014336000000000126
VERIFY 0.006607999999999947
READ 0.014176999999999662
READ 0.014210999999999974
UNPAIR 0.010548999999999698
```

A read is consistently less than 0.05s, verify is consistently less than 0.01s, pairing is consistently around 0.2s. Unpairing is consistently around 0.01s.

## This PR

```
Starting pair
PAIR 0.2088019999999995
VERIFY 0.006583999999999257
READ 0.01437699999999964
READ 0.014189000000000007
VERIFY 0.006646999999999181
READ 0.014312000000000324
READ 0.014207999999999998
VERIFY 0.007203999999999766
READ 0.014228999999999381
READ 0.014133000000000173
VERIFY 0.006713000000000413
READ 0.014281999999999684
READ 0.01422900000000027
VERIFY 0.0066030000000001365
READ 0.01424400000000059
READ 0.014348000000000027
VERIFY 0.006770000000000387
READ 0.014340999999999937
READ 0.014206000000000607
VERIFY 0.006700999999999624
READ 0.014317000000000135
READ 0.014181999999999917
VERIFY 0.0067669999999999675
READ 0.015060000000000073
READ 0.014351999999999698
VERIFY 0.0066840000000008
READ 0.01435499999999923
READ 0.014590999999999354
VERIFY 0.0066350000000001685
READ 0.014087000000000849
READ 0.014274000000000342
UNPAIR 0.011019999999999364
```

As with master, a read is consistently less than 0.05s, verify is consistently less than 0.01s, pairing is consistently around 0.2s. Unpairing is consistently around 0.01s.